+wip rename files +use case repositories rename +improve clea SOLID interface and implementation remote and assets
