--- app/src/main/java/com/rogergcc/techjobspotter/domain/IJobsPositionRepository.kt
@@ -1,13 +1,0 @@
-package com.rogergcc.techjobspotter.domain
-
-import com.rogergcc.techjobspotter.data.cloud.model.JobDto
-
-
-/**
- * Created on agosto.
- * year 2023 .
- */
-
-interface IJobsPositionRepository {
-    suspend fun geJobs(): List<JobDto>
-}
