+++ app/src/main/java/com/rogergcc/techjobspotter/domain/mappers/JobsMapperProvider.kt
@@ -1,0 +1,10 @@
+package com.rogergcc.techjobspotter.domain.mappers
+
+
+/**
+ * Created on diciembre.
+ * year 2024 .
+ */
+interface JobsMapperProvider {
+    fun getJobsMapper(): JobsMapper
+}
