+++ app/src/main/java/com/rogergcc/techjobspotter/domain/usecase/JobsPositionUseCase.kt
@@ -1,6 +1,6 @@
 package com.rogergcc.techjobspotter.domain.usecase
 
-import com.rogergcc.techjobspotter.domain.IJobsAssetsDataSource
+import com.rogergcc.techjobspotter.domain.IJobsPositions
 import com.rogergcc.techjobspotter.domain.model.JobPosition
 
 
@@ -9,8 +9,8 @@ import com.rogergcc.techjobspotter.domain.model.JobPosition
  * year 2023 .
  */
 class JobsPositionUseCase(
-    private val repository: IJobsAssetsDataSource
+    private val repository: IJobsPositions
 ) {
-    fun execute(): List<JobPosition> = repository.getJobsFromAssets()
+    suspend fun execute(): List<JobPosition> = repository.geJobs()
 
 }
