+++ app/src/main/java/com/rogergcc/techjobspotter/ui/utils/provider/ContextProvider.kt
@@ -1,0 +1,12 @@
+package com.rogergcc.techjobspotter.ui.utils.provider
+
+import android.content.Context
+
+
+/**
+ * Created on diciembre.
+ * year 2024 .
+ */
+interface ContextProvider {
+ fun getContext(): Context
+}
